telegram-bot/requirements.txt
python-telegram-bot
confuse
elasticsearch < 7.14
confluent-kafka